app/requirements.txt
--extra-index-url https://nexus.secure.amsterdam.nl/repository/pypi-hosted/simple
aniso8601==4.0.1
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
defusedxml==0.5.0
eight==0.4.2
factory-boy==2.11.1
Faker==1.0.1
flasgger==0.9.1
Flask==1.0.2
Flask-Cors==3.0.6
Flask-RESTful==0.3.6
future==0.16.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
lxml==4.2.6
MarkupSafe==1.1.0
mistune==0.8.4
pycparser==2.19
pyOpenSSL==17.5.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.20.0
signxml==2.5.2
six==1.11.0
text-unidecode==1.2
tma-saml==0.1
urllib3==1.24.1
Werkzeug==0.14.1
